refactor(dhcp): extract Lease type and accessors to lease.go
pkg/dhcp/dhcp.go fuses the DHCPv4 client, the DHCPv6 client/PD, DUID
identity management, and the lease value type into one 2,148-line file
(#6430). Begin the split by moving the Lease value type, its
LeaseRoute classless-route helper, and the read-only Leases() /
LeaseFor() accessors — plus the #1715 DNS-ownership note that documents
how Leases() feeds the daemon's reconcileDNS — verbatim into a new
sibling lease.go.

This is pure code-motion: whole types and functions move unchanged
(identical signatures and bodies); only the per-file package clause and
import block differ. The lease commit/diff lifecycle already lives in
commit.go, so lease.go is deliberately just the type plus its
accessors. A normalized multiset diff of the original file against the
concatenation of the split files shows zero logic-line delta.

Validation: go build ./pkg/dhcp/, go vet ./pkg/dhcp/ clean.
